Refactor entity and item data types to use long for IDs, enhancing consistency and performance; optimize server tick handling for mobs and items, including AI processing improvements.

+++ src/main/java/com/kneaf/core/EntityData.java
@@ -1,12 +1,12 @@
 package com.kneaf.core;
 
 public class EntityData {
-    public int id;
+    public long id;
     public double distance;
     public boolean isBlockEntity;
     public String entityType;
 
-    public EntityData(int id, double distance, boolean isBlockEntity, String entityType) {
+    public EntityData(long id, double distance, boolean isBlockEntity, String entityType) {
         this.id = id;
         this.distance = distance;
         this.isBlockEntity = isBlockEntity;
