+++ src/main/java/com/kneaf/core/ItemEntityData.java
@@ -1,14 +1,14 @@
 package com.kneaf.core;
 
 public class ItemEntityData {
-    public int id;
+    public long id;
     public int chunkX;
     public int chunkZ;
     public String itemType;
     public int count;
-    public long ageSeconds;
+    public int ageSeconds;
 
-    public ItemEntityData(int id, int chunkX, int chunkZ, String itemType, int count, long ageSeconds) {
+    public ItemEntityData(long id, int chunkX, int chunkZ, String itemType, int count, int ageSeconds) {
         this.id = id;
         this.chunkX = chunkX;
         this.chunkZ = chunkZ;
