+++ src/main/java/com/kneaf/core/MobData.java
@@ -1,12 +1,12 @@
 package com.kneaf.core;
 
 public class MobData {
-    public int id;
+    public long id;
     public double distance;
     public boolean isPassive;
     public String entityType;
 
-    public MobData(int id, double distance, boolean isPassive, String entityType) {
+    public MobData(long id, double distance, boolean isPassive, String entityType) {
         this.id = id;
         this.distance = distance;
         this.isPassive = isPassive;
